pyproject.toml
[tool.poetry]
name = "pythonproject2"
version = "0.1.0"
description = ""
authors = ["Rahul Pandey <rahulrpandey1105@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.2.12"
langchain-ollama = "^0.1.1"
langchain-community = "^0.2.11"
python-dotenv = "^1.0.1"
pypdf = "^4.3.1"
sentence-transformers = "^3.0.1"
chromadb = "^0.5.5"
langchain-chroma = "^0.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
